+++ crates/stackable-versioned-macros/src/gen/vstruct/mod.rs
@@ -1,67 +1,49 @@
-use darling::FromField;
+use std::ops::Deref;
+
 use itertools::Itertools;
 use proc_macro2::TokenStream;
-use quote::{format_ident, quote};
-use syn::{DataStruct, Error, Ident, Result};
+use quote::quote;
+use syn::{DataStruct, Error, Ident};
 
 use crate::{
-    attrs::{container::ContainerAttributes, field::FieldAttributes},
-    gen::{field::VersionedField, version::ContainerVersion},
+    attrs::common::ContainerAttributes,
+    gen::{
+        common::{format_container_from_ident, Container, ContainerVersion, VersionedContainer},
+        vstruct::field::VersionedField,
+    },
 };
+
+mod field;
 
 /// Stores individual versions of a single struct. Each version tracks field
 /// actions, which describe if the field was added, renamed or deprecated in
 /// that version. Fields which are not versioned, are included in every
 /// version of the struct.
 #[derive(Debug)]
-pub(crate) struct VersionedStruct {
-    /// The ident, or name, of the versioned struct.
-    pub(crate) ident: Ident,
+pub(crate) struct VersionedStruct(VersionedContainer<VersionedField>);
 
-    /// The name of the struct used in `From` implementations.
-    pub(crate) from_ident: Ident,
+impl Deref for VersionedStruct {
+    type Target = VersionedContainer<VersionedField>;
 
-    /// List of declared versions for this struct. Each version, except the
-    /// latest, generates a definition with appropriate fields.
-    pub(crate) versions: Vec<ContainerVersion>,
-
-    /// List of fields defined in the base struct. How, and if, a field should
-    /// generate code, is decided by the currently generated version.
-    pub(crate) fields: Vec<VersionedField>,
-
-    pub(crate) skip_from: bool,
+    fn deref(&self) -> &Self::Target {
+        &self.0
+    }
 }
 
-impl VersionedStruct {
-    pub(crate) fn new(
-        ident: Ident,
-        data: DataStruct,
-        attributes: ContainerAttributes,
-    ) -> Result<Self> {
+impl Container<DataStruct, VersionedField> for VersionedStruct {
+    fn new(ident: Ident, data: DataStruct, attributes: ContainerAttributes) -> syn::Result<Self> {
         // Convert the raw version attributes into a container version.
-        let versions = attributes
-            .versions
-            .iter()
-            .map(|v| ContainerVersion {
-                skip_from: v.skip.as_ref().map_or(false, |s| s.from.is_present()),
-                ident: format_ident!("{version}", version = v.name.to_string()),
-                deprecated: v.deprecated.is_present(),
-                inner: v.name,
-            })
-            .collect();
+        let versions: Vec<_> = (&attributes).into();
 
         // Extract the field attributes for every field from the raw token
         // stream and also validate that each field action version uses a
         // version declared by the container attribute.
-        let mut fields = Vec::new();
+        let mut items = Vec::new();
 
         for field in data.fields {
-            let attrs = FieldAttributes::from_field(&field)?;
-            attrs.validate_versions(&attributes, &field)?;
-
-            let mut versioned_field = VersionedField::new(field, attrs)?;
+            let mut versioned_field = VersionedField::new(field, &attributes)?;
             versioned_field.insert_container_versions(&versions);
-            fields.push(versioned_field);
+            items.push(versioned_field);
         }
 
         // Check for field ident collisions
@@ -69,17 +51,12 @@ impl VersionedStruct {
             // Collect the idents of all fields for a single version and then
             // ensure that all idents are unique. If they are not, return an
             // error.
-            let mut idents = Vec::new();
 
             // TODO (@Techassi): Report which field(s) use a duplicate ident and
             // also hint what can be done to fix it based on the field action /
             // status.
 
-            for field in &fields {
-                idents.push(field.get_ident(version))
-            }
-
-            if !idents.iter().all_unique() {
+            if !items.iter().map(|f| f.get_ident(version)).all_unique() {
                 return Err(Error::new(
                     ident.span(),
                     format!("struct contains renamed fields which collide with other fields in version {version}", version = version.inner),
@@ -87,27 +64,21 @@ impl VersionedStruct {
             }
         }
 
-        let from_ident = format_ident!("__sv_{ident}", ident = ident.to_string().to_lowercase());
+        let from_ident = format_container_from_ident(&ident);
 
-        Ok(Self {
+        Ok(Self(VersionedContainer {
             skip_from: attributes
                 .options
                 .skip
                 .map_or(false, |s| s.from.is_present()),
             from_ident,
             versions,
-            fields,
+            items,
             ident,
-        })
+        }))
     }
 
-    /// This generates the complete code for a single versioned struct.
-    ///
-    /// Internally, it will create a module for each declared version which
-    /// contains the struct with the appropriate fields. Additionally, it
-    /// generated `From` implementations, which enable conversion from an older
-    /// to a newer version.
-    pub(crate) fn generate_tokens(&self) -> TokenStream {
+    fn generate_tokens(&self) -> TokenStream {
         let mut token_stream = TokenStream::new();
         let mut versions = self.versions.iter().peekable();
 
@@ -117,7 +88,9 @@ impl VersionedStruct {
 
         token_stream
     }
+}
 
+impl VersionedStruct {
     fn generate_version(
         &self,
         version: &ContainerVersion,
@@ -133,14 +106,18 @@ impl VersionedStruct {
         // enable the attribute macro to be applied to a module which
         // generates versioned versions of all contained containers.
 
-        let deprecated_attr = version.deprecated.then_some(quote! {#[deprecated]});
-        let module_name = &version.ident;
+        let version_ident = &version.ident;
+
+        let deprecated_note = format!("Version {version} is deprecated", version = version_ident);
+        let deprecated_attr = version
+            .deprecated
+            .then_some(quote! {#[deprecated = #deprecated_note]});
 
         // Generate tokens for the module and the contained struct
         token_stream.extend(quote! {
             #[automatically_derived]
             #deprecated_attr
-            pub mod #module_name {
+            pub mod #version_ident {
                 pub struct #struct_name {
                     #fields
                 }
@@ -158,8 +135,8 @@ impl VersionedStruct {
     fn generate_struct_fields(&self, version: &ContainerVersion) -> TokenStream {
         let mut token_stream = TokenStream::new();
 
-        for field in &self.fields {
-            token_stream.extend(field.generate_for_struct(version));
+        for item in &self.items {
+            token_stream.extend(item.generate_for_container(version));
         }
 
         token_stream
@@ -172,9 +149,10 @@ impl VersionedStruct {
     ) -> TokenStream {
         if let Some(next_version) = next_version {
             let next_module_name = &next_version.ident;
-            let from_ident = &self.from_ident;
             let module_name = &version.ident;
-            let struct_name = &self.ident;
+
+            let from_ident = &self.from_ident;
+            let struct_ident = &self.ident;
 
             let fields = self.generate_from_fields(version, next_version, from_ident);
 
@@ -183,8 +161,8 @@ impl VersionedStruct {
             return quote! {
                 #[automatically_derived]
                 #[allow(deprecated)]
-                impl From<#module_name::#struct_name> for #next_module_name::#struct_name {
-                    fn from(#from_ident: #module_name::#struct_name) -> Self {
+                impl From<#module_name::#struct_ident> for #next_module_name::#struct_ident {
+                    fn from(#from_ident: #module_name::#struct_ident) -> Self {
                         Self {
                             #fields
                         }
@@ -204,8 +182,8 @@ impl VersionedStruct {
     ) -> TokenStream {
         let mut token_stream = TokenStream::new();
 
-        for field in &self.fields {
-            token_stream.extend(field.generate_for_from_impl(version, next_version, from_ident))
+        for item in &self.items {
+            token_stream.extend(item.generate_for_from_impl(version, next_version, from_ident))
         }
 
         token_stream